refactor: move guards
