+++ packages/pure-parse/src/guard/index.ts
@@ -1,9 +1,12 @@
-export * from './validation'
 export * from './primitives'
 export * from './json'
 export * from './types'
 export * from './literal'
-export { union } from './union'
-export { tuple } from './tuple'
-export { objectGuard } from './object'
-export { objectGuardNoJit } from './object'
+export * from './union'
+export * from './tuple'
+export * from './object'
+export * from './object'
+export * from './records'
+export * from './records'
+export * from './arrays'
+export * from './optional'
