+++ packages/pure-parse/src/guard/json.ts
@@ -1,7 +1,8 @@
-import { arrayGuard, partialRecord } from './validation'
 import { isBoolean, isNull, isNumber, isString } from './primitives'
 import { JsonValue } from '../common'
 import { union } from './union'
+import { partialRecord } from './records'
+import { arrayGuard } from './arrays'
 
 export const isJsonValue = (data: unknown): data is JsonValue =>
   union(
