+++ packages/pure-parse/src/guard/primitives.test.ts
@@ -4,7 +4,6 @@ import {
   isBigInt,
   isBoolean,
   isFunction,
-  isNonEmptyArray,
   isNull,
   isNumber,
   isObject,
@@ -13,9 +12,6 @@ import {
   isUndefined,
   isUnknown,
 } from './primitives'
-import { Equals } from '../internals'
-import { nonEmptyArray } from './validation'
-import { Infer } from '../common'
 
 describe('guards', () => {
   describe('primitives', () => {
@@ -470,56 +466,6 @@ describe('guards', () => {
       })
       it('validates functions', () => {
         expect(isFunction(() => undefined)).toEqual(true)
-      })
-    })
-  })
-
-  describe('non-empty arrays', () => {
-    describe('nonEmptyArray', () => {
-      describe('type', () => {
-        it('infers the type', () => {
-          const numberArray: number[] = [1, 2, 3]
-          if (isNonEmptyArray(numberArray)) {
-            const assertionKnownArrayType4: Equals<
-              typeof numberArray,
-              [number, ...number[]]
-            > = true
-          }
-        })
-        it('infers the exact type', () => {
-          // Number
-          const isNumberArray = nonEmptyArray(isNumber)
-          type NumberArray = Infer<typeof isNumberArray>
-          const assertionNumber1: Equals<NumberArray, [number, ...number[]]> =
-            true
-          const assertionNumber2: Equals<NumberArray, number[]> = false
-          const assertionNumber3: Equals<NumberArray, unknown[]> = false
-          const assertionNumber4: Equals<NumberArray, [unknown, ...unknown[]]> =
-            false
-          // String
-          const isStringArray = nonEmptyArray(isString)
-          type StringArray = Infer<typeof isStringArray>
-          const assertionString1: Equals<StringArray, [string, ...string[]]> =
-            true
-          const assertionString2: Equals<StringArray, string[]> = false
-          const assertionString3: Equals<StringArray, unknown[]> = false
-          const assertionString4: Equals<StringArray, [unknown, ...unknown[]]> =
-            false
-        })
-      })
-      it('validates nonempty arrays', () => {
-        expect(nonEmptyArray(isUnknown)([1])).toEqual(true)
-        expect(nonEmptyArray(isUnknown)([1, 2, 3])).toEqual(true)
-      })
-      it('invalidates empty arrays', () => {
-        expect(nonEmptyArray(isUnknown)([])).toEqual(false)
-      })
-      it('invalidates non-arrays', () => {
-        expect(nonEmptyArray(isUnknown)({})).toEqual(false)
-      })
-      it('validates each item in the array', () => {
-        expect(nonEmptyArray(isNumber)([1, 2])).toEqual(true)
-        expect(nonEmptyArray(isNumber)(['a', 'b'])).toEqual(false)
       })
     })
   })
